+++ packages/atomic/src/components/common/lazy-index.ts
@@ -2,6 +2,8 @@
 export default {
   'atomic-aria-live': async () =>
     await import('./atomic-aria-live/atomic-aria-live.js'),
+  'atomic-citation': async () =>
+    await import('./atomic-citation/atomic-citation.js'),
   'atomic-component-error': async () =>
     await import('./atomic-component-error/atomic-component-error.js'),
   'atomic-facet-date-input': async () =>
@@ -18,6 +20,12 @@ export default {
   'atomic-modal': async () => await import('./atomic-modal/atomic-modal.js'),
   'atomic-numeric-range': async () =>
     await import('./atomic-numeric-range/atomic-numeric-range.js'),
+  'atomic-result-placeholder': async () =>
+    await import('./atomic-result-placeholder/atomic-result-placeholder.js'),
+  'atomic-smart-snippet-answer': async () =>
+    await import(
+      './atomic-smart-snippet-answer/atomic-smart-snippet-answer.js'
+    ),
   'atomic-smart-snippet-collapse-wrapper': async () =>
     await import(
       './atomic-smart-snippet-collapse-wrapper/atomic-smart-snippet-collapse-wrapper.js'
@@ -25,6 +33,10 @@ export default {
   'atomic-smart-snippet-expandable-answer': async () =>
     await import(
       './atomic-smart-snippet-expandable-answer/atomic-smart-snippet-expandable-answer.js'
+    ),
+  'atomic-smart-snippet-source': async () =>
+    await import(
+      './atomic-smart-snippet-source/atomic-smart-snippet-source.js'
     ),
   'atomic-tab-bar': async () =>
     await import('./atomic-tab-bar/atomic-tab-bar.js'),