+++ packages/atomic/src/components/search/lazy-index.ts
@@ -146,6 +146,8 @@ export default {
     await import('./atomic-result-timespan/atomic-result-timespan.js'),
   'atomic-results-per-page': async () =>
     await import('./atomic-results-per-page/atomic-results-per-page.js'),
+  'atomic-search-box': async () =>
+    await import('./atomic-search-box/atomic-search-box.js'),
   'atomic-search-box-instant-results': async () =>
     await import(
       './atomic-search-box-instant-results/atomic-search-box-instant-results.js'
@@ -168,6 +170,12 @@ export default {
     await import(
       './atomic-segmented-facet-scrollable/atomic-segmented-facet-scrollable.js'
     ),
+  'atomic-smart-snippet': async () =>
+    await import('./atomic-smart-snippet/atomic-smart-snippet.js'),
+  'atomic-smart-snippet-feedback-modal': async () =>
+    await import(
+      './atomic-smart-snippet-feedback-modal/atomic-smart-snippet-feedback-modal.js'
+    ),
   'atomic-smart-snippet-suggestions': async () =>
     await import(
       './atomic-smart-snippet-suggestions/atomic-smart-snippet-suggestions.js'
@@ -182,6 +190,8 @@ export default {
   'atomic-table-element': async () =>
     await import('./atomic-table-element/atomic-table-element.js'),
   'atomic-text': async () => await import('./atomic-text/atomic-text.js'),
+  'atomic-timeframe-facet': async () =>
+    await import('./atomic-timeframe-facet/atomic-timeframe-facet.js'),
 } as Record<string, () => Promise<unknown>>;
 
 export type * from './index.js';