+++ packages/quantic/CHANGELOG.md
@@ -1,3 +1,23 @@
+## 3.35.0 (2026-01-14)
+
+* feat(quantic): Re-add the delay behaviour of tooltip on citation hover (#6775) ([52ccf0e](https://github.com/coveo/ui-kit/commits/52ccf0e)), closes [#6775](https://github.com/coveo/ui-kit/issues/6775)
+* chore: ensuring script samples are correct (#6798) ([884a3c6](https://github.com/coveo/ui-kit/commits/884a3c6)), closes [#6798](https://github.com/coveo/ui-kit/issues/6798)
+
+
+
+## <small>3.34.2 (2026-01-07)</small>
+
+* [Version Bump][skip ci]: ui-kit publish ([b2a83c4](https://github.com/coveo/ui-kit/commits/b2a83c4))
+* Add Node.js 24 (Krypton LTS) support and update CI (#6756) ([1ee69fa](https://github.com/coveo/ui-kit/commits/1ee69fa)), closes [#6756](https://github.com/coveo/ui-kit/issues/6756)
+
+
+
+## <small>3.34.1 (2026-01-06)</small>
+
+* Add Node.js 24 (Krypton LTS) support and update CI (#6756) ([1ee69fa](https://github.com/coveo/ui-kit/commits/1ee69fa)), closes [#6756](https://github.com/coveo/ui-kit/issues/6756)
+
+
+
 ## 3.34.0 (2025-12-10)
 
 * fix(quantic): better handle pushing recent results to avoid salesforce proxy extensibility errors (# ([c0b0b71](https://github.com/coveo/ui-kit/commits/c0b0b71)), closes [#6706](https://github.com/coveo/ui-kit/issues/6706)